+++ src/main/kotlin/com/ucb/edu/abc/mscompany/entity/ExchangeRateEntity.kt
@@ -3,6 +3,7 @@ package com.ucb.edu.abc.mscompany.entity
 import lombok.AllArgsConstructor
 import lombok.Data
 import lombok.NoArgsConstructor
+import java.math.BigDecimal
 import java.sql.Timestamp
 
 @Data
@@ -12,7 +13,7 @@ data class ExchangeRateEntity(
         val exchangeRateId: Int,
         var moneyName: String,
         var companyId: Int,
-        var currency: Double,
+        var currency: BigDecimal,
         var abbreviationName: String,
         var date: Timestamp
 )
